Forking mysqlx into existence
mysqlx is a Chef cookbook for setting up MySQL. It started as a fork of
the `mysql` community cookbook (sous-chefs/mysql) version 4.1.2. This
was the last version before its maintainers switched to a library
cookbook approach. This new approach using Chef resources introduced a
lot of complexity and its multi-instance setup is more or less obsolete
given that there are VMs and containers.

It'll take some work to bring the 2016 code up to modern standards but I
guess it'll be worth it!